requirements.txt
databases[sqlite]==0.4.3
fastapi==0.65.2
python-dotenv==0.18.0
requests==2.25.1
PyInquirer==1.0.3
mnemonic==0.19
qrcode[pil]==7.0
uvicorn[standard]==0.14.0
apscheduler==3.7.0
PyJWT==2.1.0
aiorpcx==0.18.7
cryptography==3.4.7
pytest==6.2.4
ecdsa==0.17.0
html2text==2020.1.16
websockets==9.1
SQLAlchemy==1.3.24
aiofiles==0.7.0
Jinja2==3.0.1
pdfkit==0.6.1
bleach==3.3.0
modules/electrum_mods/electrum-4.1.2.tar.gz
modules/electrum_mods/Electron-Cash-4.2.4.tar.gz
